# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  specialization: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  yearsOfExperience?: number | null;
  createdAt: string;
}