# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  gender: string;
  /** @nullable */
  dateOfBirth?: string | null;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  chronicConditions?: string | null;
  createdAt: string;
}