# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PrescriptionItem } from "./prescriptionItem";

export interface Prescription {
  id: number;
  patientId: number;
  doctorId: number;
  /** @nullable */
  appointmentId?: number | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  doctorName?: string | null;
  diagnosis: string;
  /** @nullable */
  notes?: string | null;
  items: PrescriptionItem[];
  createdAt: string;
}