# lib/api-zod/src/generated/types/updateLabTestRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedTrack Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLabTestRequest {
  status?: string;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  completedDate?: string | null;
}